perf: use double instead of float for edge weights
The edge weight field was stored as float but every algorithm
(Dijkstra, Louvain, ForceDirected, PageRank, etc.) operates on
doubles internally. This caused implicit float-to-double widening
on every getWeight() call in hot loops.

Changes:
- edge.java: weight field float -> double, getter/setter updated
- GraphFileParser: Float.parseFloat -> Double.parseDouble for
  precision and to avoid widening on store
- All existing float literal callers auto-widen safely (no source
  changes needed)

+++ Gvisual/src/gvisual/GraphFileParser.java
@@ -136,15 +136,15 @@ public class GraphFileParser {
                         continue;
                     }
 
-                    float weight;
+                    double weight;
                     try {
-                        weight = Float.parseFloat(parts[3]);
+                        weight = Double.parseDouble(parts[3]);
                     } catch (NumberFormatException e) {
                         LOGGER.warning("Skipping edge with invalid weight: " + line);
                         skipped++;
                         continue;
                     }
-                    if (Float.isNaN(weight) || Float.isInfinite(weight)) {
+                    if (Double.isNaN(weight) || Double.isInfinite(weight)) {
                         LOGGER.warning("Skipping edge with non-finite weight: " + line);
                         skipped++;
                         continue;
